#1015 Nested configurations
